+++ packages/design-system/src/components/SelectField/index.js
@@ -1,2 +1,6 @@
 export { default as SelectField } from './SelectField';
 export { default as Select } from './Select';
+export { default as Search } from './Search';
+export { default as SelectHead } from './SelectHead';
+export { default as SelectList } from './SelectList';
+export { default as SelectHeadItem } from './SelectHeadItem';